requirements.txt
aiosqlite==0.19.0
alabaster==0.7.16
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
apprise==1.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgi-lifespan==2.1.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Automat==22.10.0
autopep8==2.0.4
azure-common==1.1.28
azure-core==1.29.7
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
azure-storage-blob==12.19.0
azure-storage-file-share==12.15.0
azure-storage-queue==12.9.0
Babel==2.14.0
beautifulsoup4==4.12.3
black==23.12.1
bleach==6.1.0
bs4==0.0.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
constantly==23.10.4
coolname==2.2.0
croniter==2.0.1
cryptography==41.0.7
cssselect==1.2.0
dateparser==1.2.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.5.0
docker==6.1.3
docutils==0.20.1
email-validator==2.1.0.post1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.13.1
flake8==7.0.0
fqdn==1.5.1
fsspec==2023.12.2
google-auth==2.26.2
graphviz==0.20.1
greenlet==3.0.3
griffe==0.39.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
imagesize==1.4.1
incremental==22.10.0
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
isort==5.13.2
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kubernetes==29.0.0
lxml==5.1.0
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
msal==1.26.0
msal-extensions==1.1.0
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.9
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.2
orjson==3.9.12
outcome==1.3.0.post0
overrides==7.6.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parsel==1.8.1
parso==0.8.3
pathspec==0.12.1
pendulum==2.1.2
pexpect==4.9.0
platformdirs==4.1.0
portalocker==2.8.2
prefect==2.14.16
prometheus-client==0.19.0
prompt-toolkit==3.0.43
Protego==0.3.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyDispatcher==2.0.7
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.1
pyOpenSSL==23.3.0
pyperclip==1.8.2
PySocks==1.7.1
pysolr==3.9.0
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.6.2
readchar==4.0.5
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
Scrapy==2.11.0
scrapy-selenium==0.0.7
scrapy-splash==0.9.0
selenium==4.16.0
Send2Trash==1.8.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.32.0.post1
terminado==0.18.0
text-unidecode==1.3
tinycss2==1.2.1
tldextract==5.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
Twisted==22.10.0
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ujson==5.9.0
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.26.0
w3lib==2.1.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
widgetsnbextension==4.0.9
wsproto==1.2.0
zope.interface==6.1
